third_party/WebKit/LayoutTests/fast/selectors/style-sharing-shadow-expected.txt

CONSOLE WARNING: line 44: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated. See https://www.chromestatus.com/features/4668884095336448 for more details.
PASS internals.isSharingStyle(host1, host2) is true
PASS internals.isSharingStyle(querySelectorInShadow(host1, 'span'), querySelectorInShadow(host2, 'span')) is true
PASS internals.isSharingStyle(host1, host3) is false
PASS internals.isSharingStyle(host1, host4) is false
PASS internals.isSharingStyle(host1, host2) is false
PASS internals.isSharingStyle(light[0], light[1]) is true
PASS internals.isSharingStyle(light[0], shadow) is false
PASS internals.isSharingStyle(light[0], light[1]) is false
PASS successfullyParsed is true

TEST COMPLETE
item 1
item 2
A

B
